requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
pytz==2024.1
pandas==2.2.3
openpyxl==3.1.2
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.38.2
APScheduler==3.10.4
flask_apscheduler==1.12.4
python-dotenv==1.0.1
psycopg2==2.9.10
gunicorn==21.2.0
mysqlclient
PyMySQL

